pyproject.toml
[tool.poetry]
name = "social-networking-application"
version = "0.0.1"
description = ""
authors = ["Subhomoy Roy Choudhury"]
license = "MIT"
readme = "README.md"
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = "^3.11"
djangorestframework-simplejwt = "^5.3.1"
gunicorn = "^22.0.0"
mysqlclient = "2.1.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
